requirements.txt
aiohttp>=3.9.4
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna>=3.7
multidict==6.1.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
yarl==1.9.4
